+++ client/src/components/creation/Step0Method.jsx
@@ -20,7 +20,7 @@ export default function Step0Method({ onNext }) {
         <div className="wiz-cards">
           <div className="wiz-card wiz-card-disabled">
             <div className="wiz-card-media">
-              <img className="wiz-card-img" src={`${ASSETS_BASE}/PolarisCharacter01.jpg`} alt="" />
+              <img className="wiz-card-img" src={`${ASSETS_BASE}/s0_arch.webp`} alt="" />
               <div className="wiz-card-media-fade" />
             </div>
             <div className="wiz-card-body">
@@ -33,7 +33,7 @@ export default function Step0Method({ onNext }) {
 
           <div className="wiz-card wiz-card-active" onClick={onNext}>
             <div className="wiz-card-media">
-              <img className="wiz-card-img" src={`${ASSETS_BASE}/PolarisCharacter02.jpg`} alt="" />
+              <img className="wiz-card-img" src={`${ASSETS_BASE}/s0_libre.webp`} alt="" />
               <div className="wiz-card-media-fade" />
             </div>
             <div className="wiz-card-body">